TST: collect tests by method (#39976)
